Ignored client-side dependencies
This legacy version of the editor relies on implementors to download and configure client-side dependencies, such as vendor scripts. This isn't an elegant approach, but as this version is considered obsolete, we have no intention of changing it. Given that, these files should be `.gitignore`d to avoid accidentally committing third-party libraries to our repository.