Upgraded to **OnTopic Library 4.0.0**
In addition to updating to the latest version of the assemblies via **NuGet**, this also requires making quite a few updates to accommodate recent changes to the API, such as the migration from `Ignia.Topics` to `OnTopic`; the introduction of the `OnTopic.Attributes` and `OnTopic.Metadata` namespaces; and even, in some cases, the moving of `GetValue()` from `Topic` to `Attributes` (a change made quite some time ago, but which doesn't seem to have made it into the master repository for the **OnTopic Editor**). This also includes upgrading the projects to use **.NET Framework 4.8** (needed to support the **OnTopic Library**, which uses **.NET Standard 2.0**) as well as a later version of **C#** (needed to support e.g. null coalescense operators, for example). Also deleted a seemingly redundant and unnecessary copy of `/Scripts/ckeditor` in preference for `/Common/Scripts/ckeditor`, which removes a lot of clutter from the `csproj`.