Introduce `WebFormsTopicLookupService`
The legacy **OnTopic Editor** relies on a `DefaultConfiguration` attribute—previously exposed on `AttributeDescriptor`—for type specific configuration values. This was removed in **OnTopic Library 4.0.0**. Instead, it's now available on the `ConfigurableAttributeDescriptor`, which is part of **OnTopic Web Forms 4.0.0**. That said, in order for topics to be reliable created as such, the `TopicFactory.TypeLookupService` needs to be replaced with a `TypeLookupService` that's aware of the `ConfigurableAttribtueDescriptor`. This is done using the `WebFormsTopicLookupService`. (_Note:_ The **OnTopic Web Forms 4.0.0** library contains such a service as well, but there's a bug in it making it unusable for our needs; that bug will be fixes in an upcoming hotfix, but for now this should satisfy its requirements.) In addition, because the `ContentTypeDescriptor.AttributeDescriptors` collection stores values as `AttributeDescriptor`s, they need to be cast back to `ConfigurableAttributeDescriptor` anytime the code needs to operate off of the `DefaultConfiguration` or related members (e.g., `GetConfigurationValue()`).