Set default URL to `/!Admin/Topics`
This isn't strictly necessary, but makes it easier to test. The `/!Admin/Topics` path needs to be setup as a virtual directory in the `/.vs/config/applicationhost.config` on each clone of the repository. This doesn't _need_ to point to `/!Admin/Topics`, but as that's the URL we've traditionally used, it makes sense to maintain it for consistency.