Stripped out everything not absoluately necessary
The goal of the `Host` project is to be as bare bones as possible. Real implementations will obviously have more complexity and ancillary libraries; our goal is not to provide a boilerplate project they can build off of. Instead, we're demonstrating the bare minimum needed to get **OnTopic Editor** working on an **ASP.NET Web Forms** site, with an emphasis on the code and libraries that are required. As a result, I've removed a lot of dependencies that e.g. **Visual Studio** sets up out-of-the-box, or which we inherited from reference sites. It turns out, we really don't need that much!